requirements.txt
requests
numpy
scipy
Pillow